src/requirements.txt
asgiref==3.3.4
colorthief==0.2.1
Django==3.2
djangorestframework==3.12.4
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
wget==3.2
